food/report 페이지만 작업.feat : food/report 페이지만 작업.

+++ src/main/java/org/buddyguard/bodyguard/query/FoodLogWithFood.java
@@ -9,9 +9,9 @@ import java.time.LocalDate;
 @Getter
 public class FoodLogWithFood {
     // ---------------- food_logs 컬럼 ----------------
-    private long   id;
-    private long   userId;
-    private long   foodId;
+    private Integer   id;
+    private Integer   userId;
+    private Integer   foodId;
     private Double amount;
     private Double kcal;           // food_logs 쪽 kcal
     private LocalDate eatenDate;
